requirements.txt
numpy
pandas
beautifulsoup4
requests
pillow
matplotlib
pytz
geopy
tqdm
lxml
convertbng
simplekml
scikit-image
pyproj
